+++ packages/tsurlfilter/src/common/constants.ts
@@ -1,0 +1,9 @@
+export enum ErrorStatusCodes {
+    ComplexRegex = 1001,
+    RuleLimit = 1002,
+    RegexpRuleLimit = 1003,
+    RemoveparamRegexpIsNotSupported = 1004,
+    RemoveparamInversionIsNotSupported = 1005,
+}
+
+export const SEPARATOR = '|';