+++ packages/tsurlfilter/CHANGELOG.md
@@ -1,8 +1,62 @@
 # TSUrlFilter Changelog
 
+## v1.0.66
+
+### Removed
+
+- Support of $webrtc modifier
+
 
 ## 2.0.0-alpha.34
 
 ### Removed
 
 - Support of $webrtc modifier
+
+
+## v1.0.65
+
+### Changed
+
+- Updated ExtendedCss to v2.0.33
+- Updated Scriptlets to v1.7.19
+
+
+## v1.0.64
+
+### Changed
+
+- Updated Scriptlets to v1.7.14
+
+
+## v1.0.63
+
+### Changed
+
+- Updated Scriptlets to v1.7.13
+
+
+## v1.0.62
+
+### Changed
+
+- Always consider `:has()` pseudo-class as extended
+
+
+## v1.0.59
+
+### Changed
+
+- Updated ExtendedCss to v2.0.26
+
+
+## v1.0.57
+
+### Changed
+
+- Updated ExtendedCss to v2.0.24
+- Updated Scriptlets to v1.7.10
+
+### Removed
+
+- `:if()` and `:if-not()` pseudo-classes